requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
Flask==2.1.2
gunicorn==20.1.0
#heroku==0.1.4
tokenizers==0.13.2
torch==1.12.0+cpu
torchvision==0.13.0+cpu
transformers==4.24.0
nltk==3.7
numpy==1.23.1
#scipy==1.9.1
#gensim==4.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
filelock==3.8.0
huggingface-hub==0.11.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==1.5
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
six==1.16.0
smart-open==6.2.0
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.0
